Adds the ability to specify `default` parameters
This adds the ability to specify the default parameters in
`@Calendar._extracts` decorators.
If the `required` param is specified but not found, it allows to
gracefully assume a default value.

Why?
Because I have encountered that a few iCal files have missing
`VERSION` key. Rather than raising a TypeError, it makes more
sense to allow this.