[WIP] new implementation of event timespan
This implementation collects all time-related attributes of an event - begin, end/duration and precision - in a single, isolated and easy to test class.
It also targets the three possible cases for time values - timezone-aware datetimes, floating datetimes and all-day dates.
Additionally to not doing anything implicitly (neither any fancy value conversion nor fancy indirect property access),it viciously guards the validity of its internal state.
The validity of each of the 4 values representing the event time is highly depending upon the other values.
Allowing mutation opens up easy ways of breaking validity - so the implementation is actually immutable, with each modification creating a new object and requiring a full revalidation of all 4 values together.
As the times shouldn't be modified that often, having the optimization of mutability shouldn't be strictly necessary.
Additionally, moving the functionality to its own class should make testing easier and keeps this quite complicated functionality clean from other concerns.
All this makes the timespan handling easier to implement correctly, but unfortunately not easier to use.
So I also started modifying the Event class to handle the interfacing in an user-friendly way and hide the complexity of timespans.